engine/machinenativenops-auto-monitor/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "machinenativenops-auto-monitor"
version = "2.0.0"
description = "MachineNativeOps Auto Monitor - System monitoring with quantum state tracking"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MachineNativeOps Team", email = "team@machinenativeops.io"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
keywords = ["monitoring", "system", "quantum", "observability", "machinenativenops"]

requires-python = ">=3.8"
dependencies = [
    "psutil>=5.9.0",
    "prometheus-client>=0.16.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "schedule>=1.2.0",
    "kubernetes>=27.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.280",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/MachineNativeOps/machine-native-ops-aaps"
Documentation = "https://docs.machinenativeops.io"
Repository = "https://github.com/MachineNativeOps/machine-native-ops-aaps"
Issues = "https://github.com/MachineNativeOps/machine-native-ops-aaps/issues"

[project.scripts]
machinenativenops-auto-monitor = "machinenativenops_auto_monitor.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
machinenativenops_auto_monitor = ["assets/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "YTT", "S", "BLE", "B", "A", "C4", "DTZ", "T10"]
ignore = ["E501", "S301", "S311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=machinenativenops_auto_monitor --cov-report=html --cov-report=term-missing"